LineTexBucket: fix scale precision #105

+++ vtm/src/org/oscim/renderer/bucket/LineTexBucket.java
@@ -1,6 +1,6 @@
 /*
  * Copyright 2013 Hannes Janetzek
- * Copyright 2016-2017 devemux86
+ * Copyright 2016-2018 devemux86
  * Copyright 2017 Longri
  *
  * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
@@ -368,14 +368,13 @@ public final class LineTexBucket extends LineBucket {
                 GLUtils.setColor(shader.uBgColor, line.color, 1);
 
                 float pScale;
-
                 if (s >= 1) {
-                    pScale = (line.stipple * s);
-                    int cnt = (int) (pScale / line.stipple);
-                    pScale = (float) line.stipple / (cnt + 1);
+                    pScale = line.stipple * s;
+                    float cnt = pScale / line.stipple;
+                    pScale = line.stipple / (cnt + 1);
                 } else {
                     pScale = line.stipple / s;
-                    int cnt = (int) (pScale / line.stipple);
+                    float cnt = pScale / line.stipple;
                     pScale = line.stipple * cnt;
                 }
